testing/requirements.txt
psycopg2-binary
faker
python-dotenv
prometheus_client
tqdm
